Split large generated test files
