+++ src/types/src/generator/sampleArrayTarget.generator.ts
