+++ src/types/src/testGenerator.ts
@@ -2,7 +2,7 @@ import {generateFile} from './runner/manifold/generateFile'
 
 import sampleClockArray from './generator/sampleClockArray.generator'
 import sampleArrayTargetCombinable from './generator/sampleArrayTargetCombinable.generator'
-import sampleArrayTargetPlain from './generator/sampleArrayTargetPlain.generator'
+import sampleArrayTarget from './generator/sampleArrayTarget.generator'
 import sampleReturn from './generator/sampleReturn.generator'
 import sampleFilterGenerator from './generator/sampleFilter.generator'
 import sampleFn from './generator/sampleFn.generator'
@@ -10,7 +10,7 @@ import sampleFn from './generator/sampleFn.generator'
 generateFile({
   //prettier-ignore
   cases: [
-    sampleArrayTargetPlain,
+    sampleArrayTarget,
     sampleArrayTargetCombinable,
     sampleReturn,
     sampleFilterGenerator,
